Let speak animations be optional
I was following Escoria's step-by-step guide. It says that the speak animations are optional, but when I tried running the game without them, it crashed with the message "8 animation angles specified but 0 'speaks' animation(s) given". I think it makes sense to do this validation only when any speak animations are defined, i.e. their number is greater than 0.